refactor loader to use storage specific classes
