+++ packages/fireproof/src/fireproof.js
@@ -22,7 +22,7 @@ export class Fireproof {
   static storage = (name = null, opts = {}) => {
     if (name) {
       opts.name = name
-      const loader = new Loader(name, opts.loader)
+      const loader = Loader.appropriate(name, null, opts.loader)
       const existing = loader.getHeader()
       if (existing) {
         const existingConfig = JSON.parse(existing)
